requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
google-api-core==1.31.4
google-auth==1.35.0
google-cloud-bigquery==1.21.0
google-cloud-bigquery-storage==1.1.0
google-cloud-core==1.0.3
google-cloud-storage==1.18.1
google-resumable-media==0.4.1
googleapis-common-protos==1.53.0
grpcio==1.42.0
idna==3.3
numpy==1.19.5
packaging==21.3
pandas==1.1.5
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rsa==4.8
six==1.16.0
urllib3==1.26.7
